refactor setup

+++ src/actions/up-vote/up-vote.create-update.ts
@@ -1,13 +1,18 @@
 "use server";
 import { db } from "@/db/db";
 import { upVoteTable } from "@/db/schemas";
+import { validateRequest } from "@/lib/auth/validate-request";
 import { and, eq } from "drizzle-orm";
 import { generateIdFromEntropySize } from "lucia";
 import { revalidatePath } from "next/cache";
 
-export const toggleUpVote = async (setupId: string, userId: string) => {
+export const toggleUpVote = async (setupId: string) => {
+  const { user } = await validateRequest();
+
+  if (!user) { return; }
+
   const upVote = await db.query.upVoteTable.findFirst({
-    where: and(eq(upVoteTable.setupId, setupId), eq(upVoteTable.userId, userId)),
+    where: and(eq(upVoteTable.setupId, setupId), eq(upVoteTable.userId, user.id)),
   });
 
   if (upVote) { await db.delete(upVoteTable).where(eq(upVoteTable.id, upVote.id)); }
@@ -15,7 +20,7 @@ export const toggleUpVote = async (setupId: string, userId: string) => {
     await db.insert(upVoteTable).values({
       id: generateIdFromEntropySize(10),
       setupId,
-      userId
+      userId: user.id
     }).execute();
   }
